MVC och objekt oriented fixed

+++ src/main/java/controller/ShoesDescription.java
@@ -3,10 +3,8 @@ package controller;
 import javafx.collections.ObservableList;
 import javafx.scene.control.*;
 import javafx.scene.image.ImageView;
-import modell.Shoes;
+import modell.to.Shoes;
 import utils.UserLogin;
-
-import java.util.stream.Collector;
 
 /**
  * Created by Hodei Eceiza
@@ -60,7 +58,7 @@ public class ShoesDescription {
 
             //update totalPrice
             totalPriceL.setText(shoppingCart.stream().
-                    map(s -> s.getPrice() * s.getQuantity()).reduce(0, Integer::sum).toString());
+                    map(s -> s.getPrice() * s.getQuantity()).reduce(0.00, Double::sum).toString());
         });
 
         quantityS.setValueFactory(new SpinnerValueFactory.IntegerSpinnerValueFactory(0, shoesData.getQuantity()));
